Sync merge/releases/5.21 → develop

+++ common/ui/src/main/java/com/tangem/common/ui/swapStoriesScreen/SwapStoriesUM.kt
@@ -12,7 +12,7 @@ sealed class SwapStoriesUM {
 
     data class Content(
         override val stories: ImmutableList<Config>,
-        override val onClose: () -> Unit,
+        override val onClose: (Int) -> Unit,
     ) : SwapStoriesUM(), StoriesContentConfig<Content.Config> {
         override val isRestartable: Boolean = false
